+++ src/propagators/j4.jl
@@ -444,6 +444,14 @@ KeplerianElements{MeanAnomaly, Float64, Float64}:
 julia> stats
 (converged = true, iterations = 4, position_rmse = 4338.627383014923, velocity_rmse = 5.399612174061325, total_rmse = 4338.630743040775)
 ```
+
+# Extended help
+
+## Throws
+
+- `ArgumentError`: If `vjd`, `vr_i`, and `vv_i` do not have the same length, if
+    `weight_vector` does not have six elements, or if `max_iterations` is lower than 1.
+- `MeanElementsFitDivergenceError`: If the least-square iterations diverge.
 """
 function fit_j4_mean_elements(
     vjd::AbstractVector{Tjd},
@@ -590,6 +598,14 @@ KeplerianElements{MeanAnomaly, Float64, Float64}:
 julia> stats
 (converged = true, iterations = 4, position_rmse = 4338.627383014923, velocity_rmse = 5.399612174061325, total_rmse = 4338.630743040775)
 ```
+
+# Extended help
+
+## Throws
+
+- `ArgumentError`: If `vjd`, `vr_i`, and `vv_i` do not have the same length, if
+    `weight_vector` does not have six elements, or if `max_iterations` is lower than 1.
+- `MeanElementsFitDivergenceError`: If the least-square iterations diverge.
 """
 function fit_j4_mean_elements!(
     j4d::J4Propagator{Tepoch, T},