+++ src/propagators/j4osc.jl
@@ -291,6 +291,14 @@ KeplerianElements{MeanAnomaly, Float64, Float64}:
 julia> stats
 (converged = true, iterations = 4, position_rmse = 0.016876290513770688, velocity_rmse = 2.596109924485644, total_rmse = 2.596164776972825)
 ```
+
+# Extended help
+
+## Throws
+
+- `ArgumentError`: If `vjd`, `vr_i`, and `vv_i` do not have the same length, if
+    `weight_vector` does not have six elements, or if `max_iterations` is lower than 1.
+- `MeanElementsFitDivergenceError`: If the least-square iterations diverge.
 """
 function fit_j4osc_mean_elements(
     vjd::AbstractVector{Tjd},
@@ -429,6 +437,14 @@ KeplerianElements{MeanAnomaly, Float64, Float64}:
 julia> stats
 (converged = true, iterations = 4, position_rmse = 0.016876290513770688, velocity_rmse = 2.596109924485644, total_rmse = 2.596164776972825)
 ```
+
+# Extended help
+
+## Throws
+
+- `ArgumentError`: If `vjd`, `vr_i`, and `vv_i` do not have the same length, if
+    `weight_vector` does not have six elements, or if `max_iterations` is lower than 1.
+- `MeanElementsFitDivergenceError`: If the least-square iterations diverge.
 """
 function fit_j4osc_mean_elements!(
     j4oscd::J4OsculatingPropagator{Tepoch, T},