+++ test/j4.jl
@@ -509,6 +509,12 @@ end
         @test_throws ArgumentError Propagators.fit_mean_elements(
             Val(:J4), vjd, vr_i, vv_i; weight_vector = [1, 2, 3, 4, 5]
         )
+
+        # == Invalid maximum number of iterations ==========================================
+
+        @test_throws ArgumentError Propagators.fit_mean_elements(
+            Val(:J4), vjd, vr_i, vv_i; max_iterations = 0
+        )
     end
 end